Move directives, roles to their own modules
Some simple code organization. These modules are some of the few with
side effects so it's good to treat them differently.
