+++ rst2pdf/directives/aafigure.py
@@ -27,10 +27,11 @@
 
 from docutils.nodes import Element, literal_block
 from docutils.parsers.rst import directives
-from reportlab.graphics import renderPDF
 from docutils.parsers import rst
-from .opt_imports import aafigure
-from .log import log
+from reportlab.graphics import renderPDF
+
+from ..log import log
+from ..opt_imports import aafigure
 
 WARNED = False
