+++ rst2pdf/directives/code_block.py
@@ -40,7 +40,7 @@ try:
 except ImportError:
     pass
 
-from .log import log
+from ..log import log
 
 
 # Customisation
